requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
lxml>=4.9.0
schedule>=1.0.0
